src/main/java/com/iress/toyrobot/model/Direction.java
package com.iress.toyrobot.model;

public enum Direction {
    NORTH, EAST, SOUTH, WEST, NONE;

    public Direction left() {
        switch (this) {
            case NORTH: return WEST;
            case WEST: return SOUTH;
            case SOUTH: return EAST;
            case EAST: return NORTH;
            default: return NONE;
        }
    }

    public Direction right() {
        switch (this) {
            case NORTH: return EAST;
            case EAST: return SOUTH;
            case SOUTH: return WEST;
            case WEST: return NORTH;
            default: return NONE;
        }
    }
}
